\name{EnsemblRest-package}
\alias{EnsemblRest-package}
\alias{EnsemblRest}
\docType{package}
\title{
  An R interface to the Ensembl REST API
}
\description{
  TODO
}
\details{
  \tabular{ll}{
    Package: \tab EnsemblRest\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2012-10-01\cr
    License: \tab Apache License 2.0\cr
  }
}
\author{
  Tim Yates\cr
  Maintainer: Tim Yates <tyates@picr.man.ac.uk>
}
\references{
  The Ensembl REST Api: \url{http://beta.rest.ensembl.org/}
}
\keyword{ package }
